Support FIM templates and cover more models
FIM is currently a disaster. Most FIM training and infrastructure is
busted, and documentation is thin or non-existing. Is FIM just a fad?
Going in, I had thought this was the standard way people generated code
with LLMs.

Code Llama is a strange case. 7B and 13B have functioning FIM, if poor,
but 34B appears to have no FIM training despite the model card saying
otherwise. (Or maybe it's a tokenization problem?) The ommission appears
to have gone unnoticed, too. Is nobody actually doing FIM with Code
Llama?